+++ Core.Tests/Enum/TestEnum.cs
@@ -1,0 +1,12 @@
+﻿namespace Shanemat.DotNetUtils.Core.Tests.Enum;
+
+/// <summary>
+/// Represents an enumeration to use in enum tests
+/// </summary>
+internal enum TestEnum
+{
+	Unknown,
+	A,
+	B,
+	C,
+}